ios: treat warning as error

+++ test/testtimer.c
@@ -3,14 +3,14 @@
 
 typedef struct testTimerWillTriggerInOneSecondsContext {
   testItem *item;
-  int startTime;
+  unsigned int startTime;
 } testTimerWillTriggerInOneSecondsContext;
 
 static void timerTriggeredInOneSeconds(int handle) {
   int timeOffset;
   testTimerWillTriggerInOneSecondsContext *ctx;
   ctx = (testTimerWillTriggerInOneSecondsContext *)phoneGetHandleTag(handle);
-  timeOffset = abs(time(0) - ctx->startTime);
+  timeOffset = abs((int)(time(0) - ctx->startTime));
   if (timeOffset <= 2) {
     testSucceed(ctx->item);
   } else {
@@ -25,7 +25,7 @@ void testTimerWillTriggerInOneSeconds(testItem *item) {
   testTimerWillTriggerInOneSecondsContext *ctx;
   ctx = calloc(1, sizeof(testTimerWillTriggerInOneSecondsContext));
   ctx->item = item;
-  ctx->startTime = time(0);
+  ctx->startTime = (unsigned int)time(0);
   handle = phoneCreateTimer(1000, timerTriggeredInOneSeconds);
   phoneSetHandleTag(handle, ctx);
 }
